+++ src/core/mod.rs
@@ -1,2 +1,3 @@
 pub mod event;
+pub mod schedule_stream;
 pub mod transposer;
