+++ src/core/schedule_stream/mod.rs
@@ -1,3 +1,3 @@
-pub mod schedule_stream;
 pub mod realtime_stream;
+pub mod schedule_stream;
 pub mod timestamp;
