+++ src/core/transposer/mod.rs
@@ -3,4 +3,4 @@ pub mod transposer_context;
 pub mod transposer_engine;
 pub mod transposer_event;
 
-pub(self) mod transposer_engine_internal;
+pub(self) mod transposer_update;
