+++ src/main.rs
@@ -1,5 +1,5 @@
 use futures::stream::StreamExt;
-use std::{time::{Instant, Duration}};
+use std::time::{Duration, Instant};
 
 use crate::core::transposer::transposer_engine::TransposerEngine;
 use crate::example_game::{get_filtered_stream, ExampleTransposer};
@@ -9,6 +9,7 @@ use utilities::winit::WinitLoop;
 mod core;
 mod example_game;
 mod utilities;
+use crate::core::schedule_stream::schedule_stream::{ScheduleStream, ScheduleStreamExt};
 
 #[tokio::main]
 async fn main() {
@@ -17,11 +18,10 @@ async fn main() {
 
     let key_presses = get_filtered_stream(Instant::now(), receiver);
     let game: TransposerEngine<ExampleTransposer, _> = TransposerEngine::new(key_presses).await;
-    
-    let poll = game.poll(Duration::from_secs(10));
+    let stream = game.to_realtime(Instant::now());
+    let stream = stream.map(move |event| Ok(event));
     tokio::spawn(async move {
-        let poll = poll.map(move |event| Ok(event));
-        poll.forward(DebugSink::new()).await.unwrap();
+        stream.forward(DebugSink::new()).await.unwrap();
     });
 
     winit.run();
