+++ src/utilities/mod.rs
@@ -1,3 +1,3 @@
 pub mod debug_sink;
-pub mod winit;
 pub mod full_ord;
+pub mod winit;
